+++ scripts/install_shakemap_dev.sh
@@ -186,6 +186,12 @@ if [ $? -ne 0 ];then
     exit 1
 fi
 
+pip uninstall obspy -y
+echo "Installing ObsPy from GitHub..."
+if ! pip install git+https://github.com/obspy/obspy.git ; then
+    echo "Installation of ObsPy from GitHub failed."
+    exit 1
+fi
 
 if $developer; then
     echo "Installing shakemap with developer tools."